+++ internal/builder/windows.go
@@ -47,7 +47,7 @@ func (b *Windows) PreBuild(vol *volume.Volume, opts PreBuildOptions) error {
 	}
 
 	windres := b.output + ".syso"
-	err = dockerCmd(baseDockerImage, vol, []string{}, vol.TmpDirContainer(), command, opts.Verbose).Run()
+	err = dockerCmd(windowsDockerImage, vol, []string{}, vol.TmpDirContainer(), command, opts.Verbose).Run()
 	if err != nil {
 		return fmt.Errorf("Could not create the windows resource %q: %v", windres, err)
 	}
@@ -77,7 +77,7 @@ func (b *Windows) Build(vol *volume.Volume, opts BuildOptions) error {
 	}
 
 	command := goBuildCmd(output, opts)
-	err := dockerCmd(linuxDockerImage, vol, b.BuildEnv(), vol.WorkDirContainer(), command, opts.Verbose).Run()
+	err := dockerCmd(windowsDockerImage, vol, b.BuildEnv(), vol.WorkDirContainer(), command, opts.Verbose).Run()
 	if err != nil {
 		return fmt.Errorf("Could not build for %s/%s: %v", b.os, b.arch, err)
 	}