+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## Unreleased
+
 ## [1.4.0] - 2020-03-04
 - Add ability to package with an icon using fyne/cmd #14
 - Update to golang-cross:1.13.8 image (go v1.13.8) #46