+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+- Add android support #37
 
 ## [1.4.0] - 2020-03-04
 - Add ability to package with an icon using fyne/cmd #14